# packages/contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "description": "Stable error contract returned by every API failure.",
        "properties": {
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "description": "Optional structured JSON object",
            "title": "Details"
          },
          "error": {
            "description": "Stable machine-readable error code",
            "title": "Error",
            "type": "string"
          },
          "message": {
            "description": "Human-readable error summary",
            "title": "Message",
            "type": "string"
          },
          "request_id": {
            "description": "Request correlation identifier",
            "format": "uuid",
            "title": "Request Id",
            "type": "string"
          }
        },
        "required": [
          "error",
          "message",
          "request_id"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
      "ServiceResponse": {
        "properties": {
          "service": {
            "const": "fev-api",
            "title": "Service",
            "type": "string"
          },
          "status": {
            "const": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "required": [
          "service",
          "status"
        ],
        "title": "ServiceResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Flacron EnergyVerse API. Single resources are returned directly. Future list responses use `{items: [...], next_cursor: string | null}`; cursors are opaque, `null` means no more pages, and totals are not returned by default.",
    "title": "FEV API",
    "version": "0.8.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/": {
      "get": {
        "operationId": "get_root",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Unexpected server error"
          }
        },
        "summary": "Root",
        "tags": [
          "system"
        ]
      }
    },
...
  },
...
}